% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGD_bayesian_nn_predict.R
\name{SVGD_bayesian_nn_predict}
\alias{SVGD_bayesian_nn_predict}
\title{Prediction Function}
\usage{
SVGD_bayesian_nn_predict(X, theta, num_nodes, scaling_coef)
}
\arguments{
\item{X}{The covariate matrix to be predicted.}

\item{theta}{The estimated parameters in the vector format.}

\item{num_nodes}{The number of nodes in each hidden layer.}

\item{scaling_coef}{scaling coef.}
}
\value{
A vector of predicted outcomes.
}
\description{
Prediction Function
}
